Add non-generic compiledDataBinder
